requirements.txt
asgiref==3.3.1
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1

djangorestframework~=3.12.2
django-filter~=2.4.0